refactor(assertArguments): move to utilities, standardize on 'u'
