+++ lib/connection.js
@@ -13,7 +13,7 @@ var EventEmitter = require('events').EventEmitter,
     constants = require('./constants'),
     codec = require('./codec'),
     errors = require('./errors'),
-    utils = require('./utilities'),
+    u = require('./utilities'),
 
     AMQPError = require('./types/amqp_error'),
     DescribedType = require('./types/described_type'),
@@ -150,7 +150,7 @@ var EventEmitter = require('events').EventEmitter,
 function Connection(connectPolicy) {
   Connection.super_.call(this);
   var options = connectPolicy.options;
-  errors.assertArguments(options, ['containerId', 'hostname']);
+  u.assertArguments(options, ['containerId', 'hostname']);
   this.policy = connectPolicy;
   this.client = null;
   this.connected = false;
@@ -320,7 +320,7 @@ Connection.ErrorReceived = 'rxError';
  * @param sasl      If given, contains a "negotiate" method that, given address and a callback, will run through SASL negotiations.
  */
 Connection.prototype.open = function(address, sasl) {
-  errors.assertArguments(address, ['protocol', 'host', 'port']);
+  u.assertArguments(address, ['protocol', 'host', 'port']);
   this.connSM.connect(address, sasl);
 };
 
@@ -383,7 +383,7 @@ Connection.prototype._tryReceiveHeader = function(header) {
         this.connSM.headerReceived();
       }
 
-      if (utils.bufferEquals(serverVersion, constants.amqpVersion)) {
+      if (u.bufferEquals(serverVersion, constants.amqpVersion)) {
         this.connSM.validVersion();
         this.connSM.sendOpen();
       } else {
