+++ lib/errors.js
@@ -123,32 +123,6 @@ module.exports.ArgumentError = ArgumentError;
 
 
 /**
- * Convenience method to assert that a given options object contains the required arguments.
- *
- * @param options
- * @param argnames
- */
-function assertArguments(options, argnames) {
-  if (!argnames) return;
-  if (!options) throw new ArgumentError(argnames);
-  argnames.forEach(function (argname) {
-    if (!options.hasOwnProperty(argname)) {
-      throw new ArgumentError(argname);
-    }
-  });
-}
-
-module.exports.assertArguments = assertArguments;
-
-function assertArgument(arg, argname) {
-  if (arg === undefined) throw new ArgumentError(argname);
-}
-
-module.exports.assertArgument = assertArgument;
-
-
-
-/**
  * Invalid state.
  *
  * @param msg
