+++ lib/frames/attach_frame.js
@@ -168,7 +168,7 @@ function AttachFrame(options) {
     return 0;
   }
 
-  errors.assertArguments(options, ['name', 'handle', 'role']);
+  u.assertArguments(options, ['name', 'handle', 'role']);
   if (!options.source && !options.target)
     throw new errors.ArgumentError('source|target');
