+++ lib/frames/begin_frame.js
@@ -6,7 +6,6 @@ var debug = require('debug')('amqp10:framing:begin'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -92,7 +91,7 @@ function BeginFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['nextOutgoingId', 'incomingWindow', 'outgoingWindow']);
+  u.assertArguments(options, ['nextOutgoingId', 'incomingWindow', 'outgoingWindow']);
   _.defaults(this, options, {
     remoteChannel: null,
     handleMax: constants.defaultHandleMax,
