+++ lib/frames/detach_frame.js
@@ -6,7 +6,6 @@ var debug = require('debug')('amqp10:framing:detach'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -58,7 +57,7 @@ function DetachFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['handle']);
+  u.assertArguments(options, ['handle']);
   _.defaults(this, options, {
     closed: false,
     error: null
