+++ lib/frames/disposition_frame.js
@@ -6,7 +6,6 @@ var debug = require('debug')('amqp10:framing:disposition'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -98,7 +97,7 @@ function DispositionFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['role', 'first']);
+  u.assertArguments(options, ['role', 'first']);
   _.defaults(this, options, {
     last: 0,
     settled: false,
