+++ lib/frames/flow_frame.js
@@ -6,7 +6,6 @@ var debug = require('debug')('amqp10:framing:flow'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -152,7 +151,7 @@ function FlowFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['incomingWindow', 'nextOutgoingId', 'outgoingWindow']);
+  u.assertArguments(options, ['incomingWindow', 'nextOutgoingId', 'outgoingWindow']);
   _.defaults(this, options, {
     nextIncomingId: null,
     handle: null,
