+++ lib/frames/open_frame.js
@@ -6,7 +6,6 @@ var debug = require('debug')('amqp10:framing:open'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -155,7 +154,7 @@ function OpenFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['containerId', 'hostname']);
+  u.assertArguments(options, ['containerId', 'hostname']);
   this.id = options.containerId;
   _.defaults(this, options, {
     maxFrameSize: constants.defaultMaxFrameSize,
@@ -178,7 +177,7 @@ OpenFrame.Descriptor = {
 
 OpenFrame.prototype._getPerformative = function() {
   var self = this;
-  errors.assertArguments(self, ['id', 'hostname']);
+  u.assertArguments(self, ['id', 'hostname']);
   return new DescribedType(OpenFrame.Descriptor.code, {
     id: self.id,
     hostname: self.hostname,
