+++ lib/frames/sasl_frame.js
@@ -216,7 +216,7 @@ function SaslInit(options) {
     this.initialResponse = up.get(options, 1);
     this.hostname = up.get(options, 2);
   } else {
-    errors.assertArguments(options, ['mechanism']);
+    u.assertArguments(options, ['mechanism']);
     this.mechanism = options.mechanism;
     this.initialResponse = options.initialResponse;
     this.hostname = options.hostname;
@@ -265,7 +265,7 @@ function SaslChallenge(options) {
   } else if (options instanceof Buffer) {
     this.challenge = options;
   } else {
-    errors.assertArguments(options, ['challenge']);
+    u.assertArguments(options, ['challenge']);
     this.challenge = options.challenge;
   }
 }
@@ -310,7 +310,7 @@ function SaslResponse(options) {
   } else if (options instanceof Buffer) {
     this.response = options;
   } else {
-    errors.assertArguments(options, ['response']);
+    u.assertArguments(options, ['response']);
     this.response = options.response;
   }
 }
@@ -385,7 +385,7 @@ function SaslOutcome(options) {
     this.additionalData = up.get(options, 1);
     this.details = constants.saslOutcomes[this.code];
   } else {
-    errors.assertArguments(options, ['code']);
+    u.assertArguments(options, ['code']);
     this.code = options.code;
     this.additionalData = options.additionalData;
   }
