+++ lib/frames/transfer_frame.js
@@ -7,7 +7,6 @@ var debug = require('debug')('amqp10:framing:transfer'),
     _ = require('lodash'),
 
     constants = require('../constants'),
-    errors = require('../errors'),
     u = require('../utilities'),
     up = u.payload,
 
@@ -194,7 +193,7 @@ function TransferFrame(options) {
     return;
   }
 
-  errors.assertArguments(options, ['handle']);
+  u.assertArguments(options, ['handle']);
   _.defaults(this, options, {
     deliveryId: null,
     deliveryTag: null,
