+++ lib/sasl.js
@@ -30,7 +30,7 @@ function Sasl() {
 }
 
 Sasl.prototype.negotiate = function(connection, credentials, done) {
-  errors.assertArguments(credentials, ['user', 'pass']);
+  u.assertArguments(credentials, ['user', 'pass']);
   this.connection = connection;
   this.credentials = credentials;
   this.callback = done;
