+++ lib/session.js
@@ -178,7 +178,7 @@ Session.DispositionReceived = 'disposition';
 
 Session.prototype.begin = function(sessionPolicy) {
   var sessionParams = u.deepMerge(sessionPolicy.options);
-  errors.assertArguments(sessionParams, ['nextOutgoingId', 'incomingWindow', 'outgoingWindow']);
+  u.assertArguments(sessionParams, ['nextOutgoingId', 'incomingWindow', 'outgoingWindow']);
 
   this.policy = sessionPolicy;
   this.channel = this.connection.associateSession(this);
