+++ lib/types/delivery_state.js
@@ -5,6 +5,7 @@ var Int64 = require('node-int64'),
 
     constants = require('../constants'),
     errors = require('../errors'),
+    u = require('../utilities'),
 
     AMQPError = require('./amqp_error'),
     DescribedType = require('./described_type'),
@@ -21,7 +22,7 @@ module.exports.DeliveryState = DeliveryState;
 
 function Received(options) {
   Received.super_.call(this, Received.Descriptor.code);
-  errors.assertArguments(options, ['sectionNumber', 'sectionOffset']);
+  u.assertArguments(options, ['sectionNumber', 'sectionOffset']);
   this.sectionNumber = options.sectionNumber;
   this.sectionOffset = options.sectionOffset;
 }
