+++ lib/utilities.js
@@ -196,3 +196,27 @@ function coerce(val, T) {
 }
 
 module.exports.coerce = coerce;
+
+/**
+ * Convenience method to assert that a given options object contains the required arguments.
+ *
+ * @param options
+ * @param argnames
+ */
+function assertArguments(options, argnames) {
+  if (!argnames) return;
+  if (!options) throw new errors.ArgumentError(argnames);
+  argnames.forEach(function (argname) {
+    if (!options.hasOwnProperty(argname)) {
+      throw new errors.ArgumentError(argname);
+    }
+  });
+}
+
+module.exports.assertArguments = assertArguments;
+
+function assertArgument(arg, argname) {
+  if (arg === undefined) throw new errors.ArgumentError(argname);
+}
+
+module.exports.assertArgument = assertArgument;
